# Improve performance of manipulating the attributes.