--- a/editor/src/benchmarks.benchmark.ts
+++ b/editor/src/benchmarks.benchmark.ts
@@ -1,7 +1,9 @@
 import { benchmarkBuildTree } from './core/shared/element-path-tree.benchmark'
 import { benchmarkElementPathFunction } from './core/shared/element-path.benchmark'
+import { benchmarkAttributes } from './core/shared/jsx-attributes.benchmark'
 import { benchmarkPropertyPathFunction } from './core/shared/property-path.benchmark'
 
-await benchmarkBuildTree()
-await benchmarkElementPathFunction()
-await benchmarkPropertyPathFunction()
+//await benchmarkBuildTree()
+//await benchmarkElementPathFunction()
+//await benchmarkPropertyPathFunction()
+await benchmarkAttributes()