--- a/editor/src/core/model/element-template-utils.ts
+++ b/editor/src/core/model/element-template-utils.ts
@@ -57,7 +57,7 @@ import {
 import { assertNever, fastForEach } from '../shared/utils'
 import { getComponentsFromTopLevelElements, isSceneAgainstImports } from './project-file-utils'
 import { getStoryboardElementPath } from './scene-utils'
-import { getJSXAttributeAtPath, GetJSXAttributeResult } from '../shared/jsx-attributes'
+import { getJSXAttributesAtPath, GetJSXAttributeResult } from '../shared/jsx-attributes'
 import { forceNotNull, optionalMap } from '../shared/optional-utils'
 import {
   ConditionalCase,
@@ -793,10 +793,13 @@ export function componentHonoursPropsPosition(component: UtopiaJSXComponent): bo
   } else {
     const rootElement = component.rootElement
     if (isJSXElement(rootElement)) {
-      const leftStyleAttr = getJSXAttributeAtPath(rootElement.props, PP.create('style', 'left'))
-      const topStyleAttr = getJSXAttributeAtPath(rootElement.props, PP.create('style', 'top'))
-      const rightStyleAttr = getJSXAttributeAtPath(rootElement.props, PP.create('style', 'right'))
-      const bottomStyleAttr = getJSXAttributeAtPath(rootElement.props, PP.create('style', 'bottom'))
+      const leftStyleAttr = getJSXAttributesAtPath(rootElement.props, PP.create('style', 'left'))
+      const topStyleAttr = getJSXAttributesAtPath(rootElement.props, PP.create('style', 'top'))
+      const rightStyleAttr = getJSXAttributesAtPath(rootElement.props, PP.create('style', 'right'))
+      const bottomStyleAttr = getJSXAttributesAtPath(
+        rootElement.props,
+        PP.create('style', 'bottom'),
+      )
       return (
         ((propertyComesFromPropsStyle(component.param, leftStyleAttr, 'left') ||
           propertyComesFromPropsStyle(component.param, rightStyleAttr, 'right')) &&
@@ -816,8 +819,11 @@ export function componentHonoursPropsSize(component: UtopiaJSXComponent): boolea
   } else {
     const rootElement = component.rootElement
     if (isJSXElement(rootElement)) {
-      const widthStyleAttr = getJSXAttributeAtPath(rootElement.props, PP.create('style', 'width'))
-      const heightStyleAttr = getJSXAttributeAtPath(rootElement.props, PP.create('style', 'height'))
+      const widthStyleAttr = getJSXAttributesAtPath(rootElement.props, PP.create('style', 'width'))
+      const heightStyleAttr = getJSXAttributesAtPath(
+        rootElement.props,
+        PP.create('style', 'height'),
+      )
       return (
         (propertyComesFromPropsStyle(component.param, widthStyleAttr, 'width') &&
           propertyComesFromPropsStyle(component.param, heightStyleAttr, 'height')) ||
@@ -833,7 +839,7 @@ export function propsStyleIsSpreadInto(propsParam: Param, attributes: JSXAttribu
   const boundParam = propsParam.boundParam
   switch (boundParam.type) {
     case 'REGULAR_PARAM': {
-      const styleProp = getJSXAttributeAtPath(attributes, PP.create('style'))
+      const styleProp = getJSXAttributesAtPath(attributes, PP.create('style'))
       const styleAttribute = styleProp.attribute
       switch (styleAttribute.type) {
         case 'ATTRIBUTE_NOT_FOUND':
@@ -876,7 +882,7 @@ export function propsStyleIsSpreadInto(propsParam: Param, attributes: JSXAttribu
             // This is the aliased name or if there's no alias the field name.
             const propertyToLookFor = partBoundParam.paramName
 
-            const styleProp = getJSXAttributeAtPath(attributes, PP.create('style'))
+            const styleProp = getJSXAttributesAtPath(attributes, PP.create('style'))
             const styleAttribute = styleProp.attribute
             switch (styleAttribute.type) {
               case 'ATTRIBUTE_NOT_FOUND':