--- a/editor/src/core/shared/jsx-attributes.benchmark.ts
+++ b/editor/src/core/shared/jsx-attributes.benchmark.ts
@@ -1,0 +1,156 @@
+import * as Benny from 'benny'
+import * as PP from './property-path'
+import { getJSXAttributesAtPath, jsxAttributesToProps } from './jsx-attributes'
+import {
+  emptyComments,
+  jsExpressionNestedObject,
+  jsExpressionOtherJavaScript,
+  jsExpressionValue,
+  jsxAttributeNestedArraySimple,
+  jsxAttributeNestedObjectSimple,
+  JSXAttributes,
+  jsxAttributesFromMap,
+  jsxPropertyAssignment,
+  jsxSpreadAssignment,
+} from './element-template'
+import { NO_OP } from './utils'
+
+function sampleJsxAttributes(): JSXAttributes {
+  return jsxAttributesFromMap({
+    style: jsExpressionNestedObject(
+      [
+        jsxSpreadAssignment(
+          jsExpressionValue({ first: 1, second: 2 }, emptyComments),
+          emptyComments,
+        ),
+        jsxPropertyAssignment(
+          'backgroundColor',
+          jsExpressionValue('red', emptyComments),
+          emptyComments,
+          emptyComments,
+        ),
+        jsxPropertyAssignment(
+          'shadow',
+          jsExpressionOtherJavaScript(
+            'props.someShadow',
+            'return props.someShadow;',
+            ['props'],
+            null,
+            {},
+          ),
+          emptyComments,
+          emptyComments,
+        ),
+        jsxPropertyAssignment(
+          'border',
+          jsExpressionValue('1px solid green', emptyComments),
+          emptyComments,
+          emptyComments,
+        ),
+        jsxPropertyAssignment(
+          'boxShadow',
+          jsExpressionValue('0 0 0 1px blue', emptyComments),
+          emptyComments,
+          emptyComments,
+        ),
+        jsxPropertyAssignment(
+          'top',
+          jsExpressionValue(0, emptyComments),
+          emptyComments,
+          emptyComments,
+        ),
+        jsxPropertyAssignment(
+          'left',
+          jsExpressionValue(50, emptyComments),
+          emptyComments,
+          emptyComments,
+        ),
+        jsxPropertyAssignment(
+          'height',
+          jsExpressionValue(150, emptyComments),
+          emptyComments,
+          emptyComments,
+        ),
+        jsxPropertyAssignment(
+          'width',
+          jsExpressionValue(200, emptyComments),
+          emptyComments,
+          emptyComments,
+        ),
+      ],
+      emptyComments,
+    ),
+    objectWithArray: jsxAttributeNestedObjectSimple(
+      jsxAttributesFromMap({
+        array: jsExpressionValue([0, 1, 2], emptyComments),
+      }),
+      emptyComments,
+    ),
+    objectWithNestedArray: jsxAttributeNestedObjectSimple(
+      jsxAttributesFromMap({
+        array: jsxAttributeNestedArraySimple([
+          jsExpressionValue(0, emptyComments),
+          jsExpressionValue(1, emptyComments),
+          jsExpressionValue(2, emptyComments),
+        ]),
+      }),
+      emptyComments,
+    ),
+    doggo: jsExpressionOtherJavaScript('props.hello', 'return props.hello;', ['props'], null, {}),
+    objectValue: jsExpressionValue(
+      {
+        deep: {
+          object: {
+            path: 'yes',
+          },
+        },
+      },
+      emptyComments,
+    ),
+    otherJs: jsExpressionOtherJavaScript('true ? 10 : 5', 'return true ? 10 : 5', [], null, {}),
+    'data-uid': jsExpressionValue('aaa', emptyComments),
+  })
+}
+
+export async function benchmarkAttributes(): Promise<void> {
+  await Benny.suite(
+    'get an attribute',
+    Benny.add('getJSXAttributesAtPath', () => {
+      const propertyPath = PP.create('style', 'width')
+      const attributes = sampleJsxAttributes()
+      return () => {
+        const result = getJSXAttributesAtPath(attributes, propertyPath)
+        switch (result.attribute.type) {
+          case 'ATTRIBUTE_NOT_FOUND':
+            throw new Error(`getJSXAttributeAtPath failed: ATTRIBUTE_NOT_FOUND`)
+          case 'PART_OF_ATTRIBUTE_VALUE':
+            throw new Error(`getJSXAttributeAtPath failed: PART_OF_ATTRIBUTE_VALUE`)
+          default:
+          // Do nothing, we got a value.
+        }
+      }
+    }),
+    Benny.cycle(),
+    Benny.complete(),
+    Benny.save({ file: 'get an attribute', details: true }),
+  )
+  await Benny.suite(
+    'convert attributes into props',
+    Benny.add('jsxAttributesToProps', () => {
+      const inScope = {
+        someValue: 1,
+        anotherValue: 'another',
+        props: {
+          hello: 'Hello World',
+        },
+      }
+      const attributes = sampleJsxAttributes()
+      return () => {
+        jsxAttributesToProps('test.js', inScope, attributes, NO_OP)
+      }
+    }),
+    Benny.cycle(),
+    Benny.complete(),
+    Benny.save({ file: 'convert attributes to props', details: true }),
+  )
+}