… editor/src/components/canvas/commands/convert-to-absolute-command.spec.tsx
… editor/src/components/editor/actions/__snapshots__/actions.spec.tsx.snap
… editor/src/components/editor/actions/actions.spec.tsx
… editor/src/components/inspector/common/css-utils.spec.ts
… editor/src/components/inspector/common/property-controls.spec.browser2.tsx
… editor/src/core/layout/layout-utils.spec.ts
… editor/src/core/model/jsx-attributes.spec.ts
… editor/src/core/property-controls/property-control-values.spec.ts
… editor/src/core/shared/jsx-attributes.ts
… editor/src/core/shared/property-path.benchmark.ts
… editor/src/core/workers/parser-printer/parser-printer.ts