**Problem:**
Lots of the logic of the editor manipulates the attributes in some way, so any performance we can gain there should benefit the whole editor.

**Fix:**
Most of the improvements came from eliminating where possible the manipulation of property path values, using instead the array of elements backing the path and an index value into that array.

Other than this there are mostly tweaks to for loops, avoiding reversing arrays by winding an index backwards and changing another by inverting the contents and stopping once one part applies.

**Results:**
```
getJSXAttributesAtPath:
8,164,725 ops/s, ±0.62% -> 33,450,101 ops/s, ±0.47%

jsxAttributesToProps:
251,281 ops/s, ±1.13% -> 405,767 ops/s, ±1.31%

setJSXValueAtPath:
234,550 ops/s, ±0.62% -> 280,677 ops/s, ±0.12%
```

**Commit Details:**
- Renamed `getJSXAttributeAtPath` to `getJSXAttributesAtPath`.
- Made `JSXPropertyAssignment` and `JSXAttributesEntry` more permissive in allowing keys of `string | number`.
- Refactored `getJSXAttribute` to not reverse the array of attributes.
- Added `getJSXAttributesAtPathParts`, replacing the innards of `getJSXAttributesAtPath`, to avoid creating multiple instances of `ElementPath`.
-  Added `benchmarkAttributes`.
- Don't rely on the behaviour of `Array.concat` in jsxSimpleAttributeToValue` and `jsxAttributeToValue`.
- Replace a `map` and `sequenceEither` with a `traverseEither`.
- Sped up object creation in `jsxAttributeToValue`.
- Made `deleteJSXAttribute` more permissive with what can be passed to the `key` parameter.
- Refactored `resolveDefinedElsewhere` to not need to execute the entire body of the for loop.
- Put the majority of the logic of `setJSXValueAtPath` into `setJSXValueAtPathParts`.
- Put the majority of the logic of `setJSXValueInAttributeAtPath` into `setJSXValueInAttributeAtPathParts`.
- Tweaked `benchmarkPropertyPathFunction` to clear the property path cache after each suite.
- Refactored `getInsertionSubjectWrapperConditionalFalseBranch` to use `getSimpleAttributeAtPath`, so that it can handle more variations of `JSXAttributes`.